CSS rules define the styles for the TrendCard component and its child elements:
.TrendCard:

display: flex: The TrendCard component is displayed as a flex container.
flex-direction: column: Child elements of the TrendCard are arranged in a column layout.
gap: 1rem: Adds a gap of 1rem between the child elements of the TrendCard.
background-color: var(--cardColor): Sets the background color of the TrendCard to a custom variable --cardColor.
padding: 1rem: Applies padding of 1rem to the TrendCard.
border-radius: 1rem: Rounds the corners of the TrendCard with a border-radius of 1rem.
padding-left: 2rem: Adds additional left padding of 2rem to the TrendCard.
.trend:

display: flex: The .trend elements are displayed as flex containers.
flex-direction: column: Child elements of the .trend are arranged in a column layout.
gap: 0.5rem: Adds a gap of 0.5rem between the child elements of the .trend.
.trend > span
(1):

Targets the first span element within each .trend element.
font-weight: bold: Sets the font weight to bold for the first span element, likely representing the trend name.
.trend > span
(2):

Targets the second span element within each .trend element.
font-size: 13px: Sets the font size to 13 pixels for the second span element, likely representing the number of shares for the trend.
These styles ensure that the TrendCard component and its child elements are visually consistent and appropriately spaced, providing a cohesive and readable presentation of trending topics.